style: 기존 뉴스 히스토리 테스트 제거, 명칭 변경
